requirements.txt
pyserial
python-osc